♻️ [Refactoring]: deployment::executor モジュール分離 + execute() match 分解
- src/component/deployment/executor.rs に ComponentDeployment 構造体本体と impl を抽出
- execute() の match を deploy_skill / deploy_command / deploy_agent / deploy_instruction / deploy_hook の 5 メソッドに分解
- フィールド可視性を pub(super) に整え、兄弟モジュール (hook_deploy.rs) からアクセス可能に
- src/component/deployment/executor_test.rs に Execute(6) + Accessor(1) + Command 変換(3) + Agent 変換(4) = 14 件を移動
- hook_deploy.rs の import を super::ComponentDeployment → super::executor::ComponentDeployment に差し替え
- deployment.rs ルートは Builder 定義のみが残る状態に
- 既存 N=59 件 green 維持
